+++ js/script.js
@@ -30,8 +30,6 @@
 		 */
 		displayParsedEmail: function(fileName) {
 			const parserUrl = OC.generateUrl(OCA.FilesEmlViewer.PreviewEml._baseUrl + '/emlparse?eml_file={file}', {file: fileName});
-			const printerUrl = OC.generateUrl(OCA.FilesEmlViewer.PreviewEml._baseUrl + '/emlparse?eml_file={file}&print=', {file: fileName});
-			const pdfUrl = OC.generateUrl(OCA.FilesEmlViewer.PreviewEml._baseUrl + '/pdf?eml_file={file}', {file: fileName});
 			$.ajax({
 				async: false,
 				method: 'GET',
@@ -44,9 +42,6 @@
 							$("#email-text-content").toggleClass("fade-out");
 							$('#toggle-text-content').toggleText('Show raw content', 'Hide raw content');
 						});
-
-						$("#make-pdf").attr('href',pdfUrl);
-						$("#printer-friendly").attr('href',printerUrl);
 					}
 				},
 				error: function(response) {